+++ mitmproxy/addons/export.py
@@ -24,14 +24,14 @@ def cleanup_request(f: flow.Flow) -> http.Request:
     return request
 
 
-def pop_headers(request: http.Request) -> http.Request:
-    # Remove some headers that are redundant for curl/httpie export
+def pop_headers(request: http.Request) -> None:
+    """Remove some headers that are redundant for curl/httpie export."""
     request.headers.pop("content-length", None)
+
     if request.headers.get("host", "") == request.host:
         request.headers.pop("host")
     if request.headers.get(":authority", "") == request.host:
         request.headers.pop(":authority")
-    return request
 
 
 def cleanup_response(f: flow.Flow) -> http.Response:
@@ -62,7 +62,8 @@ def request_content_for_console(request: http.Request) -> str:
 
 def curl_command(f: flow.Flow) -> str:
     request = cleanup_request(f)
-    request = pop_headers(request)
+    pop_headers(request)
+
     args = ["curl"]
 
     server_addr = f.server_conn.peername[0] if f.server_conn.peername else None
@@ -83,6 +84,12 @@ def curl_command(f: flow.Flow) -> str:
             args += ["-H", f"{k}: {v}"]
 
     if request.method != "GET":
+        if not request.content:
+            # curl will not calculate content-length if there is no content
+            # some server/verb combinations require content-length headers
+            # (ex. nginx and POST)
+            args += ["-H", "content-length: 0"]
+
         args += ["-X", request.method]
 
     args.append(request.pretty_url)
@@ -95,7 +102,7 @@ def curl_command(f: flow.Flow) -> str:
 
 def httpie_command(f: flow.Flow) -> str:
     request = cleanup_request(f)
-    request = pop_headers(request)
+    pop_headers(request)
 
     # TODO: Once https://github.com/httpie/httpie/issues/414 is implemented, we
     # should ensure we always connect to the IP address specified in the flow,